+++ quantization/tests/test_sse.rs
@@ -31,12 +31,12 @@ mod tests {
         .unwrap();
         let query_u8 = encoded.encode_query(&query);
 
-        let indexes = (0..vectors_count).collect::<Vec<_>>();
+        let indexes = (0..vectors_count as u32).collect::<Vec<_>>();
         let mut scores = vec![0.0; vectors_count];
         encoded.score_points_sse(&query_u8, &indexes, &mut scores);
 
         for i in 0..vectors_count {
-            let score = encoded.score_point_sse(&query_u8, i);
+            let score = encoded.score_point_sse(&query_u8, i as u32);
             let score2 = scores[i];
             let orginal_score = dot_similarity(&query, &vector_data[i]);
             assert!((score - orginal_score).abs() < error);
@@ -68,12 +68,12 @@ mod tests {
         .unwrap();
         let query_u8 = encoded.encode_query(&query);
 
-        let indexes = (0..vectors_count).collect::<Vec<_>>();
+        let indexes = (0..vectors_count as u32).collect::<Vec<_>>();
         let mut scores = vec![0.0; vectors_count];
         encoded.score_points_sse(&query_u8, &indexes, &mut scores);
 
         for i in 0..vectors_count {
-            let score = encoded.score_point_sse(&query_u8, i);
+            let score = encoded.score_point_sse(&query_u8, i as u32);
             let score2 = scores[i];
             let orginal_score = l2_similarity(&query, &vector_data[i]);
             assert!((score - orginal_score).abs() < error);
